+++ crates/goose-providers/src/formats.rs
@@ -1,4 +1,5 @@
 pub mod anthropic;
+pub mod databricks;
 pub mod ollama;
 pub mod openai;
 pub mod openai_responses;
