+++ crates/goose-providers/src/lib.rs
@@ -3,6 +3,9 @@ pub mod api_client;
 pub mod base;
 pub mod canonical;
 pub mod conversation;
+pub mod databricks;
+pub mod databricks_auth;
+pub mod databricks_v2;
 pub mod declarative;
 pub mod errors;
 pub mod formats;
