+++ crates/goose/src/providers/formats/mod.rs
@@ -3,7 +3,9 @@ pub mod anthropic {
 }
 #[cfg(feature = "aws-providers")]
 pub mod bedrock;
-pub mod databricks;
+pub mod databricks {
+    pub use goose_providers::formats::databricks::*;
+}
 pub mod gcpvertexai;
 pub mod google;
 pub mod openrouter;
