+++ crates/goose/src/providers/mod.rs
@@ -31,7 +31,9 @@ pub mod cursor_agent;
 pub mod custom_provider_config;
 pub mod databricks;
 pub mod databricks_auth;
+pub mod databricks_def;
 pub mod databricks_v2;
+pub mod databricks_v2_def;
 pub mod formats;
 mod gcpauth;
 pub mod gcpvertexai;
